pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcbench"
version = "0.1.0"
description = "Modular benchmarks for matrix completion algorithms."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["numpy>=1.24", "PyYAML>=6.0"]

[project.scripts]
mcbench = "mcbench.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
